[feat] pop seed + ordered list prop

+++ app/Http/Controllers/Admin/PropertyController.php
@@ -12,6 +12,9 @@ use App\Models\Sponsor;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
+use Illuminate\Support\Carbon;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class PropertyController extends Controller
 {
@@ -54,8 +57,8 @@ class PropertyController extends Controller
         $form_data = $request->validated();
         $form_data['slug'] = Property::generateSlug($form_data['title']);
 
-        if ($request->hasFile('cover_image')) {
-            $form_data['cover_image'] = Storage::put('cover_image', $form_data['cover_image']);
+        if ($request->hasFile('cover_image') && $request->file('cover_image')->isValid()) {
+            $form_data['cover_image'] = Storage::put('cover_image', $request->file('cover_image'));
         } else {
             $form_data['cover_image'] = 'https://placehold.co/600x400?text=Cover+Image';
         }
@@ -81,11 +84,35 @@ class PropertyController extends Controller
      * @param  \App\Models\Property  $property
      * @return \Illuminate\Http\Response
      */
-    public function show(Property $property)
+    public function show(Request $request, Property $property)
     {
         // $userId = Auth::id();
-
         // $properties = Property::where('user_id', $userId)->get();
+        // Ottieni l'indirizzo IP del visitatore
+        $ipAddress = $request->ip();
+        // Cerca una vista esistente per questa combinazione di IP e proprietà
+        $view = DB::table('views')
+            ->where('ip_address', $ipAddress)
+            ->where('property_id', $property->id)
+            ->first();
+
+        if (!$view || Carbon::parse($view->updated_at)->diffInMinutes(now()) >= 5) {
+            if ($view) {
+                // Aggiorna solo il timestamp di updated_at se esiste già
+                DB::table('views')
+                    ->where('id', $view->id)
+                    ->update(['updated_at' => now()]);
+            } else {
+                // Se la vista non esiste, crea una nuova riga
+                DB::table('views')->insert([
+                    'ip_address' => $ipAddress,
+                    'property_id' => $property->id,
+                    'created_at' => now(),
+                    'updated_at' => now(),
+                ]);
+            }
+        }
+
         $sponsors = Sponsor::all();
         $services = Service::all();
         $images = Image::all();
@@ -120,12 +147,13 @@ class PropertyController extends Controller
         $form_data = $request->validated();
         $form_data['slug'] = Property::generateSlug($form_data['title']);
 
-        if ($request->hasFile('cover_image')) {
+        if ($request->hasFile('cover_image') && $request->file('cover_image')->isValid()) {
+
             if (Str::startsWith($property->cover_image, 'https') === false) {
                 Storage::delete($property->cover_image);
             }
 
-            $form_data['cover_image'] = Storage::put('cover_image', $form_data['cover_image']);
+            $form_data['cover_image'] = Storage::put('cover_image', $request->file('cover_image'));
         }
 
         $property->update($form_data);
@@ -151,12 +179,11 @@ class PropertyController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Property $property)
-    { {
-            if (Str::startsWith($property->cover_image, 'https') === false) {
-                Storage::delete($property->cover_image);
-            }
-            $property->delete();
-            return redirect()->route("admin.properties.index")->with("success", "Annuncio cancellato");
+    {
+        if (Str::startsWith($property->cover_image, 'https') === false) {
+            Storage::delete($property->cover_image);
         }
+        $property->delete();
+        return redirect()->route("admin.properties.index")->with("success", "Annuncio cancellato");
     }
 }
